add outline of required view classes for bundle api endpoints
