+++ backend/marketplace/views/__init__.py
@@ -1,0 +1,2 @@
+from bundles import *
+from seller import *
